refactor: 태그 관련 컴포넌트 구조화 및 ArticleList 개선
- SelectTag 컴포넌트를 ui/tag 디렉토리로 이동
- ArticleList에서 새로운 TagList 컴포넌트로 태그 렌더링 로직 분리
- 태그 렌더링의 재사용성 및 모듈성 향상
